documents: update files permissions
* Updates files permissions using a CLI command and an input CSV file.
* Moves existing CLI commands for documents.
* Closes #147.
