+++ sonar/modules/documents/jsonschemas/documents/document-v1.0.0_src.json
@@ -1036,6 +1036,17 @@
             "type": "string",
             "minLength": 1,
             "pattern": "^https?://"
+          },
+          "restricted": {
+            "title": "Restricted",
+            "type": "string",
+            "minLength": 1
+          },
+          "embargo_date": {
+            "title": "Embargo date",
+            "type": "string",
+            "minLength": 1,
+            "format": "date"
           }
         },
         "required": [
