+++ tests/ui/documents/data/invalid.csv
@@ -1,0 +1,2 @@
+"id_bibrec";"status"
+14278;"status:PA16JU"
